// src/enums/route-ids-enum.ts
export enum RouteIdsEnum {
  HOME = 'HOME',
  EMPLOYEES = 'EMPLOYEES',
  ATTENDANCE_LOGS = 'ATTENDANCE_LOGS',
  NATIONALITIES = 'NATIONALITIES',
  CITIES = 'CITIES',
  REGIONS = 'REGIONS',
  PERMISSION_REASONS = 'PERMISSION_REASONS',
  GENERAL_SETTINGS = 'GENERAL_SETTINGS',
  DASHBOARD = 'DASHBOARD',
  HOLIDAYS = 'HOLIDAYS',
  PERMISSIONS = 'PERMISSIONS',
  WORK_SHIFT_LIST = ' WORK_SHIFT_LIST',
  WORK_SHIFT_ASSIGNMENT = ' WORK_SHIFT_ASSIGNMENT',
  WORK_SHIFT_TEMP = 'WORK_SHIFT_TEMP',
  WORK_SHIFTS = 'WORK_SHIFTS',
  DEPARTMENTS = 'DEPARTMENTS',
  NOTIFICATIONS = 'NOTIFICATIONS',
  BLACKLIST = 'BLACKLIST',
  PRESENCE_INQUIRIES = 'PRESENCE_INQUIRIES',
  LIMITED_TIME_PERMISSION = 'LIMITED_TIME_PERMISSION',
  WORK_MISSION = 'WORK_MISSION',
  VISIT_REQUEST = 'VISIT_REQUEST',
  ACCESS_LOCATIONS = 'ACCESS_LOCATIONS',
  DEVICES_CONFIGURATION = 'DEVICES_CONFIGURATION',
  ATTENDANCE_REPORT = 'ATTENDANCE_REPORT',
  ATTENDANCE_REPORT_PROCESSING = 'ATTENDANCE_REPORT_PROCESSING',
  LEAVE_TYPES = 'LEAVE_TYPES',
  LEAVES_BALANCES = 'LEAVES_BALANCES',
}